requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
open-clip-torch==2.20.1
moviepy==1.0.3
Werkzeug==2.3.6